+++ src/doc/index.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import Syscall from './syscall';
 import Vfs from './vfs';
 import BuildinFunction from './buildinFunction';
+import Macro from './macro';
 import Probe from './probe';
 
 let syscall: vscode.CompletionItem[] = [];
@@ -13,9 +14,11 @@ for (const i of Syscall) {
     let name = i.name;
     name = name.replace(/syscall\./, '');
     let action = new vscode.CompletionItem(name);
+    action.kind = vscode.CompletionItemKind.Property;
     action.documentation = i.doc;
     syscall.push(action);
     let action_return = new vscode.CompletionItem(name + '.return');
+    action_return.kind = vscode.CompletionItemKind.Property;
     action_return.documentation = i.doc;
     syscall.push(action_return);
 }
@@ -24,9 +27,11 @@ for (const i of Vfs) {
     let name = i.name;
     name = name.replace(/vfs\./, '');
     let action = new vscode.CompletionItem(name);
+    action.kind = vscode.CompletionItemKind.Property;
     action.documentation = i.doc;
     vfs.push(action);
     let action_return = new vscode.CompletionItem(name + '.return');
+    action_return.kind = vscode.CompletionItemKind.Property;
     action_return.documentation = i.doc;
     vfs.push(action_return);
 }
@@ -34,6 +39,15 @@ for (const i of Vfs) {
 for (const i of BuildinFunction) {
     let name = i.name;
     let action = new vscode.CompletionItem(name);
+    action.kind = vscode.CompletionItemKind.Function;
+    action.documentation = new vscode.MarkdownString(i.doc);
+    buildinFunction.push(action);
+}
+
+for (const i of Macro) {
+    let name = i.name;
+    let action = new vscode.CompletionItem(name);
+    action.kind = vscode.CompletionItemKind.Function;
     action.documentation = new vscode.MarkdownString(i.doc);
     buildinFunction.push(action);
 }
@@ -44,6 +58,7 @@ for (const i of Probe) {
     let name = fullname.replace(type + '.', '');
     probe[type] = probe[type] || [];
     let action = new vscode.CompletionItem(name);
+    action.kind = vscode.CompletionItemKind.Property;
     action.documentation = new vscode.MarkdownString(i.doc);
     probe[type].push(action);
 }